+++ src/main/java/com/metropolitan/it355/controller/GostController.java
@@ -3,10 +3,13 @@ package com.metropolitan.it355.controller;
 import com.metropolitan.it355.entity.Gost;
 import com.metropolitan.it355.services.GostService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @AllArgsConstructor
 @RestController
@@ -21,8 +24,11 @@ public class GostController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Gost> findById(@PathVariable int id) {
-        return ResponseEntity.ok(gostService.getById(id));
+    public ResponseEntity<?> findById(@PathVariable int id) {
+        if (gostService.getById(id).isPresent()){
+            return ResponseEntity.ok(gostService.getById(id).get());
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
 
     @PostMapping
@@ -36,10 +42,14 @@ public class GostController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Gost> delete(@PathVariable int id) {
-        Gost gost = gostService.getById(id);
-        gostService.delete(id);
-        return ResponseEntity.ok(gost);
+    public ResponseEntity<?> delete(@PathVariable int id) {
+        Optional<?> gost = gostService.getById(id);
+
+        if (gost.isPresent()) {
+            gostService.delete(id);
+            return ResponseEntity.ok(gost);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
