+++ src/main/java/com/metropolitan/it355/controller/RezervacijaController.java
@@ -3,10 +3,13 @@ package com.metropolitan.it355.controller;
 import com.metropolitan.it355.entity.Rezervacija;
 import com.metropolitan.it355.services.RezervacijaService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @AllArgsConstructor
 @RestController
@@ -21,8 +24,11 @@ public class RezervacijaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Rezervacija> findById(@PathVariable int id) {
-        return ResponseEntity.ok(rezervacijaService.getById(id));
+    public ResponseEntity<?> findById(@PathVariable int id) {
+        if (rezervacijaService.getById(id).isPresent()){
+            return ResponseEntity.ok(rezervacijaService.getById(id).get());
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
 
     @GetMapping("/gost/{id}")
@@ -51,10 +57,14 @@ public class RezervacijaController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Rezervacija> delete(@PathVariable int id) {
-        Rezervacija rezervacija = rezervacijaService.getById(id);
-        rezervacijaService.delete(id);
-        return ResponseEntity.ok(rezervacija);
+    public ResponseEntity<?> delete(@PathVariable int id) {
+        Optional<?> rezervacija = rezervacijaService.getById(id);
+
+        if (rezervacija.isPresent()) {
+            rezervacijaService.delete(id);
+            return ResponseEntity.ok(rezervacija);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
 
 }
