+++ src/main/java/com/metropolitan/it355/controller/TransportController.java
@@ -3,10 +3,13 @@ package com.metropolitan.it355.controller;
 import com.metropolitan.it355.entity.Transport;
 import com.metropolitan.it355.services.TransportService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @AllArgsConstructor
 @RestController
@@ -21,8 +24,11 @@ public class TransportController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Transport> findById(@PathVariable int id) {
-        return ResponseEntity.ok(transportService.getById(id));
+    public ResponseEntity<?> findById(@PathVariable int id) {
+        if (transportService.getById(id).isPresent()){
+            return ResponseEntity.ok(transportService.getById(id).get());
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
 
     @PostMapping
@@ -36,10 +42,13 @@ public class TransportController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Transport> delete(@PathVariable int id) {
-        Transport transport = transportService.getById(id);
-        transportService.delete(id);
-        return ResponseEntity.ok(transport);
+    public ResponseEntity<?> delete(@PathVariable int id) {
+        Optional<?> transport = transportService.getById(id);
+        if (transport.isPresent()) {
+            transportService.delete(id);
+            return ResponseEntity.ok(transport);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Map.of("message", "Not found"));
     }
